added sound change and bitrate

+++ gstreamer_interface/include/SessionManager.h
@@ -11,41 +11,41 @@
 class SessionManager : public QObject {
 private:
     Q_OBJECT
-	QHash<QPair<QHostAddress, QString>, std::shared_ptr<Session>> live_sessions;  // map<session_id, session>
-        // void handleException(GstreamerError error);	 // Provides flowless application work after gstreamer errors.
+    QHash<QPair<QHostAddress, QString>, std::shared_ptr<Session>> live_sessions;  // map<session_id, session>
+    // void handleException(GstreamerError error);	 // Provides flowless application work after gstreamer errors.
 
-        void startThread(Session* session, const QString session_type);
+    void startThread(Session* session, const QString session_type);
 
-    public:
-        SessionManager();
+public:
+    SessionManager();
 
-	~SessionManager();
+    ~SessionManager();
 
 public slots:
-	void onStartVideoSession(const QHostAddress ip_address);
+    void onStartVideoSession(const QHostAddress ip_address);
 
-	void onStartAudioSession(const QHostAddress ip_address);
+    void onStartAudioSession(const QHostAddress ip_address);
 
-	void onKillVideoSession(const QHostAddress ip_address);
+    void onKillVideoSession(const QHostAddress ip_address);
 
-	void onKillAudioSession(const QHostAddress ip_address);
+    void onKillAudioSession(const QHostAddress ip_address);
 
-	void onStartReceivingSession(const QHostAddress ip_address, const QString session_type);
+    void onStartReceivingSession(const QHostAddress ip_address, const QString session_type);
 
-	void onKillVideoReciver(const QHostAddress ip_address);
+    void onKillVideoReciver(const QHostAddress ip_address);
 
     void onKillAudioReciver(const QHostAddress ip_address);
 
     void onReceivedPorts(const QHostAddress ip_address, qint32 video_port, qint32 audio_port);
 
-    void onSetVolume(const QHostAddress ip_address, const float volume);
+    void onSetVolume(const QHostAddress ip_address, const int volume);
 
     void onSetBitrate(const QHostAddress ip_address, const int bitrate);
 
 signals:
     void sendErrorOccured(const QString error_string);
 
-    void setVolume(const float volume);
+    void setVolume(const int volume);
 
     void setBitrate(const int bitrate);
